// apps/portal/src/ldap/ldap.interface.ts
export type Scope = 'base' | 'one' | 'sub';
// ...
interface GroupSearchFilterFunction {
  /**
   * Construct a group search filter from user object
   *
   * @param user The user retrieved and authenticated from LDAP
   */
  (user: any): string;
}

export interface LdapModuleOptions extends ClientOptions {
  /**
   * Logger function
   */
  logger?: Logger;

  /**
   * Admin connection DN, e.g. uid=myapp,ou=users,dc=example,dc=org.
   * If not given at all, admin client is not bound. Giving empty
   * string may result in anonymous bind when allowed.
   *
   * Note: Not passed to ldapjs, it would bind automatically
   */
  bindDN: string;
  /**
   * Password for bindDN
   */
  bindCredentials: string;
  /**
   * Property of the LDAP user object to use when binding to verify
   * the password. E.g. name, email. Default: dn
   */
  bindProperty?: 'dn';

  /**
   * The base DN from which to search for users by username.
   * E.g. ou=users,dc=example,dc=org
   */
  searchBase: string;
  /**
   * LDAP search filter with which to find a user by username, e.g.
   * (uid={{username}}). Use the literal {{username}} to have the
   * given username interpolated in for the LDAP search.
   */
  searchFilter: string;
  /**
   * Scope of the search. Default: 'sub'
   */
  searchScope?: Scope;
  /**
   * Array of attributes to fetch from LDAP server. Default: all
   */
  searchAttributes?: string[];

  /**
   * The base DN from which to search for synchronization.
   * E.g. ou=users,dc=example,dc=org
   */
  searchBaseAllUsers: string;
  /**
   * LDAP search filter with synchronization.
   */
  searchFilterAllUsers: string;
  /**
   * Scope of the search. Default: 'sub'
   */
  searchScopeAllUsers?: Scope;
  /**
   * Array of attributes to fetch from LDAP server. Default: all
   */
  searchAttributesAllUsers?: string[];

  /**
   * The base DN from which to search for groups. If defined,
   * also groupSearchFilter must be defined for the search to work.
   */
  groupSearchBase?: string;
  /**
   * LDAP search filter for groups. Place literal {{dn}} in the filter
   * to have it replaced by the property defined with `groupDnProperty`
   * of the found user object. Optionally you can also assign a
   * function instead. The found user is passed to the function and it
   * should return a valid search filter for the group search.
   */
  groupSearchFilter?: string | GroupSearchFilterFunction;
  /**
   * Scope of the search. Default: sub
   */
  groupSearchScope?: Scope;
  /**
   * Array of attributes to fetch from LDAP server. Default: all
   */
  groupSearchAttributes?: string[];

  /**
   * The property of user object to use in '{{dn}}' interpolation of
   * groupSearchFilter. Default: 'dn'
   */
  groupDnProperty?: string;

  /**
   * Set to true to add property '_raw' containing the original buffers
   * to the returned user object. Useful when you need to handle binary
   * attributes
   */
  includeRaw?: boolean;

  timeLimit?: number;
  sizeLimit?: number;

  /**
   * If true, then up to 100 credentials at a time will be cached for
   * 5 minutes.
   */
  cache?: boolean;
}

export interface LdapOptionsFactory {
  createLdapOptions(): Promise<LdapModuleOptions> | LdapModuleOptions;
}

export interface LdapModuleAsyncOptions extends Pick<ModuleMetadata, 'imports'> {
  useExisting?: Type<LdapOptionsFactory>;
  useClass?: Type<LdapOptionsFactory>;
  useFactory?: (...args: any[]) => Promise<LdapModuleOptions> | LdapModuleOptions;
  inject?: any[];
}

export const ldapADattributes = [
  'thumbnailPhoto;binary',
  'objectGUID;binary',
  'objectSid;binary',
  'c',
  'cn',
  'co',
  'codePage',
  'comment',
  'company',
  'countryCode',
  'department',
  'description',
  'displayName',
  'distinguishedName',
  'dn',
  'employeeID',
  'flags',
  'givenName',
  'l',
  'mail',
  'memberOf',
  'middleName',
  'manager',
  'mobile',
  'name',
  'objectCategory',
  'objectClass',
  'otherMailbox',
  'otherTelephone',
  'postalCode',
  'primaryGroupID',
  'sAMAccountName',
  'sAMAccountType',
  'sn',
  'st',
  'streetAddress',
  'title',
  'wWWHomePage',
  'userAccountControl',
];